B-4 Add testCreateInstanceFromDDMMYYY, testCreateInstanceFromSerial and testCreateInstanceFromJavaDate
- All tests are added to SerialDateTest from BobsSerialDateTest so it is removed
