+++ src/test/java/org/jfree/date/SerialDateTest.java
@@ -59,6 +59,7 @@ import java.io.ObjectInput;
 import java.io.ObjectInputStream;
 import java.io.ObjectOutput;
 import java.io.ObjectOutputStream;
+import java.util.GregorianCalendar;
 
 import junit.framework.Test;
 import junit.framework.TestCase;
@@ -735,4 +736,23 @@ public class SerialDateTest extends TestCase {
         //}
     }
 
+    public void testCreateInstanceFromDDMMYYY() throws Exception {
+        SerialDate date = createInstance(1, JANUARY, 1900);
+        assertEquals(1,date.getDayOfMonth());
+        assertEquals(JANUARY,date.getMonth());
+        assertEquals(1900,date.getYYYY());
+        assertEquals(2,date.toSerial());
+    }
+
+    public void testCreateInstanceFromSerial() throws Exception {
+        assertEquals(d(1, JANUARY, 1900), createInstance(2));
+        assertEquals(d(1, JANUARY, 1901), createInstance(367));
+    }
+
+    public void testCreateInstanceFromJavaDate() throws Exception {
+        assertEquals(d(1, JANUARY, 1900),
+                createInstance(new GregorianCalendar(1900,0,1).getTime()));
+        assertEquals(d(1, JANUARY, 2006),
+                createInstance(new GregorianCalendar(2006,0,1).getTime()));
+    }
 }
